+++ src/bika/lims/content/abstractanalysis.py
@@ -51,8 +51,10 @@ from Products.Archetypes.Field import StringField
 from Products.Archetypes.references import HoldingReference
 from Products.Archetypes.Schema import Schema
 from Products.CMFCore.permissions import View
+from senaite.core.api import dtime
 from senaite.core.browser.fields.datetime import DateTimeField
 from senaite.core.i18n import translate as t
+from senaite.core.i18n import get_dt_format
 from senaite.core.permissions import FieldEditAnalysisResult
 from senaite.core.permissions import ViewResults
 from six import string_types
@@ -529,6 +531,20 @@ class AbstractAnalysis(AbstractBaseAnalysis):
         # Ensure result integrity regards to None, empty and 0 values
         val = str("" if not value and value != 0 else value).strip()
 
+        # Check if a date/time result
+        result_type = self.getResultType()
+        if result_type in ["date", "datetime"]:
+            # convert to datetime
+            dt = dtime.to_dt(val)
+            # make it TZ-naive to prevent undesired shifts
+            dt = dt.replace(tzinfo=None) if dt else None
+            # store as ISO format for easy handling
+            with_time = result_type == "datetime"
+            fmt = "%Y-%m-%d %H:%M:%S" if with_time else "%Y-%m-%d"
+            val = dtime.date_to_string(dt, fmt=fmt)
+            self.getField("Result").set(self, val)
+            return
+
         # Check if an string result is expected
         string_result = self.getStringResult()
 
@@ -905,8 +921,21 @@ class AbstractAnalysis(AbstractBaseAnalysis):
             except (ValueError, TypeError):
                 pass
 
-        # If string result, return without any formatting
-        if self.getStringResult():
+        result_type = self.getResultType()
+
+        # If date result, apply the format for current locale, without TZ
+        if result_type in ["date", "datetime"]:
+            # convert to datetime
+            dt = dtime.to_dt(result)
+            # make TZ-naive to prevent undesired shifts
+            dt = dt.replace(tzinfo=None) if dt else None
+            # apply format set for current locale, but without localizing
+            fmt = get_dt_format(result_type)
+            result = dtime.date_to_string(dt, fmt)
+            return cgi.escape(result) if html else result
+
+        # If string-like result, return without any formatting
+        if result_type in ["string", "text"]:
             if html:
                 result = cgi.escape(result)
                 result = result.replace("\n", "<br/>")