+++ src/senaite/core/api/dtime.py
@@ -39,10 +39,20 @@ from DateTime.DateTime import DateError
 from DateTime.DateTime import DateTimeError
 from DateTime.DateTime import SyntaxError
 from DateTime.DateTime import TimeError
+from senaite.core.i18n import get_month_name
+from senaite.core.i18n import get_weekday_name
 from zope.i18n import translate
 
 
 RX_GMT = r"^(\bGMT\b|)([+-]?)(\d{1,2})"
+
+YMD_REGEX = (
+    r"^"
+    r"((?P<y>(\d+))y){0,1}\s*"  # years
+    r"((?P<m>(\d+))m){0,1}\s*"  # months
+    r"((?P<d>(\d+))d){0,1}\s*"  # days
+    r"((?P<h>(\d+))h){0,1}\s*"  # hours
+)
 
 _marker = object()
 
@@ -428,6 +438,25 @@ def to_iso_format(dt):
     return None
 
 
+def to_msgstr(fmt):
+    """Converts to the format used by the TranslationServiceTool, that supports
+    the following directives, each used as a variable in msgstr:
+       ${A} ${a} ${B} ${b} ${H} ${I} ${m} ${d} ${M} ${p} ${S} ${Y} ${y} ${Z}
+
+    For example: "${d}-${m}-${Y} ${H}:${M}"
+
+    Note the following directives from C standard are NOT supported:
+        %w %f %z %j %U %W %c %x %X
+    """
+    return re.sub(r"%([aAdbBmyYHIpMSZ])", r"${\1}", fmt)
+
+
+def to_C1989(fmt):
+    """Converts to a format code that adheres to the C standard (1989 version)
+    """
+    return re.sub(r"\${([aAwdbBmyYHIpMSfzZjUWcxX])}", r"%\1", fmt)
+
+
 def date_to_string(dt, fmt="%Y-%m-%d", default=""):
     """Format the date to string
     """
@@ -440,31 +469,15 @@ def date_to_string(dt, fmt="%Y-%m-%d", default=""):
     if isinstance(dt, six.string_types):
         dt = to_DT(dt)
 
+    # convert the format to C Standard (1989 version) just in case
+    fmt_c1989 = to_C1989(fmt)
+
     try:
-        return dt.strftime(fmt)
+        return dt.strftime(fmt_c1989) or default
     except ValueError:
-        #  Fix ValueError: year=1111 is before 1900;
-        #  the datetime strftime() methods require year >= 1900
-
-        # convert format string to be something like "${Y}-${m}-${d}"
-        new_fmt = ""
-        var = False
-        for x in fmt:
-            if x == "%":
-                var = True
-                new_fmt += "${"
-                continue
-            if var:
-                new_fmt += x
-                new_fmt += "}"
-                var = False
-            else:
-                new_fmt += x
-
-        def pad(val):
-            """Add a zero if val is a single digit
-            """
-            return "{:0>2}".format(val)
+        # The datetime strftime() function requires year >= 1900
+        # Use safe_substitute instead
+        new_fmt = to_msgstr(fmt)
 
         # Manually extract relevant date and time parts
         dt = to_DT(dt)
@@ -473,14 +486,28 @@ def date_to_string(dt, fmt="%Y-%m-%d", default=""):
             "y": dt.yy(),
             "m": dt.mm(),
             "d": dt.dd(),
-            "H": pad(dt.h_24()),
-            "I": pad(dt.h_12()),
-            "M": pad(dt.minute()),
+            "H": "{:0>2}".format(dt.h_24()),
+            "I": "{:0>2}".format(dt.h_12()),
+            "M": "{:0>2}".format(dt.minute()),
             "p": dt.ampm().upper(),
-            "S": dt.second(),
+            "S": "{:0>2}".format(dt.second()),
+            "Z": get_timezone(dt),
+            "b": get_month_name(dt.month(), abbr=True),
+            "B": get_month_name(dt.month()),
         }
 
-        return Template(new_fmt).safe_substitute(data)
+        if "${w}" in new_fmt:
+            data["w"] = int(dt.strftime("%w"))
+
+        if "${a}" in new_fmt.lower():
+            # Weekday name ${A} or abbreviation (${a})
+            weekday = int(dt.strftime("%w"))
+            data.update({
+                "a": get_weekday_name(weekday, abbr=True),
+                "A": get_weekday_name(weekday),
+            })
+
+        return Template(new_fmt).safe_substitute(data) or default
 
 
 def to_localized_time(dt, long_format=None, time_only=None,
@@ -532,7 +559,7 @@ def to_localized_time(dt, long_format=None, time_only=None,
     return time_str
 
 
-def get_relative_delta(dt1, dt2=None):
+def get_relativedelta(dt1, dt2=None):
     """Calculates the relative delta between two dates or datetimes
 
     If `dt2` is None, the current datetime is used.
@@ -568,6 +595,10 @@ def get_relative_delta(dt1, dt2=None):
         dt2 = dt2.replace(tzinfo=tzinfo)
 
     return relativedelta(dt2, dt1)
+
+
+# BBB
+get_relative_delta = get_relativedelta
 
 
 def timedelta_to_dict(value, default=_marker):
@@ -632,3 +663,172 @@ def to_timedelta(value, default=_marker):
         minutes=to_int(value.get('minutes', 0), 0),
         seconds=to_int(value.get('seconds', 0), 0)
     )
+
+
+def is_ymd(ymd):
+    """Returns whether the string represents a duration of time in ymd format
+
+    :param ymd: supposedly ymd string to evaluate
+    :type ymd: str
+    :returns: True if a valid duration in ymd format
+    :rtype: bool
+    """
+    if not is_str(ymd):
+        return False
+    try:
+        to_ymd(ymd)
+    except (TypeError, ValueError):
+        return False
+    return True
+
+
+def ymd(years=0, months=0, days=0, hours=0):
+    """Returns a string representing a time duration in ymd format
+
+    :param years: years
+    :type years: int
+    :param months: months
+    :type months: int
+    :param days: days
+    :type days: int
+    :param hours: hours
+    :type hours: int
+    :param default: fall-back value to return as default
+    :returns: a string that represents a duration in ymd format
+    :rtype: str
+    """
+    # create the relativedelta to take shifts into account
+    delta = relativedelta(years=years, months=months, days=days, hours=hours)
+    delta = delta.normalized()
+
+    # apply ymd format, with zeros omitted
+    values = [delta.years, delta.months, delta.days, delta.hours]
+    values = map(lambda val: str(abs(val)), values)
+    value = filter(lambda it: int(it[0]), zip(values, "ymdh"))
+    value = " ".join(map("".join, value))
+
+    # return a compliant ymd
+    return value or "0d"
+
+
+def to_ymd(duration, with_hours=False, default=_marker):
+    """Returns the given duration in ymd format
+
+    If default is _marker, either a TypeError or ValueError is raised if
+    the duration is not valid or cannot be converted to ymd format
+
+    :param duration: duration to be converted to a ymd format
+    :type duration: str/relativedelta
+    :param with_hours: whether hours have to be included in the output or not
+    :type with_hours: bool
+    :param default: fall-back value to return as default
+    :returns: a string that represents a duration in ymd format
+    :rtype: str
+    """
+    try:
+        delta = to_relativedelta(duration)
+        hours = delta.hours if with_hours else 0
+        return ymd(delta.years, delta.months, delta.days, hours)
+    except (TypeError, ValueError) as e:
+        if default is _marker:
+            raise e
+        return default
+
+
+def to_relativedelta(duration, normalized=True):
+    """Returns the given duration as a relativedelta
+
+    If default is _marker, either a TypeError or ValueError is raised if
+    the duration is not valid or cannot be converted to relativedelta format
+
+    :param duration: duration as length of time
+    :type delta: str/relativedelta/tuple/list
+    :param normalized: whether relative attributes are represented as integers
+    :type normalized: bool
+    :returns: a duration in time represented as a relativedelta object
+    :rtype: relativedelta
+    """
+    if isinstance(duration, relativedelta):
+        return duration.normalized() if normalized else duration
+
+    if isinstance(duration, (tuple, list)):
+        # convert to tuple of (years, months, days, hours) to int values
+        keys = ["years", "months", "days", "hours"]
+        values = [to_int(value, 0) for value in duration]
+        values = dict(zip(keys, values))
+        delta = relativedelta(**values)
+        return delta.normalized() if normalized else delta
+
+    if not is_str(duration):
+        raise TypeError("{} is not supported".format(repr(duration)))
+
+    # to lowercase and remove leading and trailing spaces
+    raw_ymd = duration.lower().strip()
+
+    # extract the years, months and days
+    matches = re.search(YMD_REGEX, raw_ymd)
+    values = [matches.group(v) for v in "ymdh"]
+
+    # if all values are None, assume the ymd format was not valid
+    nones = [value is None for value in values]
+    if all(nones):
+        raise ValueError("Not a valid ymd: {}".format(repr(duration)))
+
+    # replace Nones with zeros and calculate everything with a relativedelta
+    keys = ["years", "months", "days", "hours"]
+    values = [to_int(value, 0) for value in values]
+    values = dict(zip(keys, values))
+    delta = relativedelta(**values)
+    return delta.normalized() if normalized else delta
+
+
+def get_since_date(duration, dt=None, default=_marker):
+    """Returns the relative date when the event started
+
+    If dt is None, the function uses the current date time as the date from
+    which the relative date is calculated.
+
+    When delta is not a valid period and default value is _marker, a TypeError
+    or ValueError is raised. Otherwise, it returns the default value converted
+    to datetime (or None if it cannot be converted)
+
+    :param duration: duration as length of time
+    :type duration: str/relativedelta/tuple/list
+    :param dt: date from which the since date has to be calculated
+    :type dt: string/DateTime/datetime/date
+    :param default: fall-back date-like value to return as default
+    :returns: the calculated relative date
+    :rtype: tuple
+    """
+    try:
+        delta = to_relativedelta(duration)
+    except (TypeError, ValueError) as e:
+        if default is _marker:
+            raise e
+        return to_dt(default)
+
+    # datetime from which the relative date has to be calculated
+    dt = to_dt(dt) if dt else now()
+
+    # calculate the date when everything started
+    return dt - delta
+
+
+def get_ymd(dt1, dt2=None, with_hours=False):
+    """Calculates the relative delta between two dates or datetimes and
+    returns the duration in ymd format
+
+    If `dt2` is None, the current datetime is used.
+
+    :param dt1: the first date/time to compare
+    :type dt1: string/date/datetime/DateTime
+    :param dt2: the second date/time to compare
+    :type dt2: string/date/datetime/DateTime
+    :returns: interval of time in ymd format (e.g. "2y4m")
+    :rtype: str
+    """
+    try:
+        delta = get_relativedelta(dt1, dt2=dt2)
+        return to_ymd(delta, with_hours=with_hours)
+    except (ValueError, TypeError):
+        return None